pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "cronboard"
authors = [
  { name = "Antonio Rodriguez", email = "kontakt@antoniorodriguez.no" },
]
description = "A terminal-based dashboard for managing cron jobs"
version = "0.2.0"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
  "bcrypt>=5.0.0",
  "cron-descriptor>=2.0.6",
  "croniter>=6.0.0",
  "dt-croniter>=6.0.1",
  "paramiko>=4.0.0",
  "pytest>=8.4.2",
  "python-crontab>=3.3.0",
  "textual>=6.2.1",
  "textual-autocomplete>=4.0.6",
  "textual-dev>=1.7.0",
]

[project.urls]
Repository = "https://github.com/antoniorodr/cronboard"
Issues = "https://github.com/antoniorodr/cronboard/issues"

[project.scripts]
cronboard = "cronboard.app:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.pytest.ini_options]
pythonpath = [".", "src"]

[tool.setuptools.package-data]
"cronboard" = ["static/css/*.tcss"]
